Move exchange components to pages/exchange
