+++ app/routes.js
@@ -6,7 +6,7 @@ import { Route, Switch } from 'react-router';
 import Layout from './containers/Layout';
 import HomePage from './pages/PortfolioPage';
 import WalletPage from './pages/WalletPage';
-import { ExchangePage } from './pages/ExchangePage';
+import { ExchangePage } from './pages/exchange/ExchangePage';
 import TransactionPage from './pages/TransactionPage';
 import SettingsPage from './pages/SettingsPage';
 import AboutPage from './pages/AboutPage';
